Revert "Все Id сущностей переведены с int на long (нужно для авторизации через VK ID)"
This reverts commit f9799ae1855e2c217aa4d1cb18e889b7c9ee2397.

+++ Kventin.Backend/Kventin.Domain/Domain/Announcement.cs
@@ -20,7 +20,7 @@ namespace Kventin.DataAccess.Domain
         /// <summary>
         /// Id автора
         /// </summary>
-        public long AuthorId { get; set; }
+        public int AuthorId { get; set; }
 
         /// <summary>
         /// Автор
