+++ Kventin.Backend/Kventin.Domain/Domain/Mark.cs
@@ -21,7 +21,7 @@ namespace Kventin.DataAccess.Domain
         /// <summary>
         /// Id ученика
         /// </summary>
-        public long StudentId { get; set; }
+        public int StudentId { get; set; }
 
         /// <summary>
         /// Ученик
@@ -31,7 +31,7 @@ namespace Kventin.DataAccess.Domain
         /// <summary>
         /// Id преподавателя
         /// </summary>
-        public long TeacherId { get; set; }
+        public int TeacherId { get; set; }
 
         /// <summary>
         /// Преподаватель
@@ -41,7 +41,7 @@ namespace Kventin.DataAccess.Domain
         /// <summary>
         /// Id Занятия
         /// </summary>
-        public long? LessonId { get; set; }
+        public int? LessonId { get; set; }
 
         /// <summary>
         /// Занятие
@@ -51,7 +51,7 @@ namespace Kventin.DataAccess.Domain
         /// <summary>
         /// Id задания
         /// </summary>
-        public long? ExerciseId { get; set; }
+        public int? ExerciseId { get; set; }
 
         /// <summary>
         /// Задание
