+++ Kventin.Backend/Kventin.Services/Dtos/Marks/StudentMarksDto.cs
@@ -5,7 +5,7 @@ namespace Kventin.Services.Dtos.Marks
     public class StudentMarksDto
     {
         [Required]
-        public long StudentId { get; set; }
+        public int StudentId { get; set; }
 
         public List<MarkShortInfoDto> Marks { get; set; } = [];
     }
