+++ Kventin.Backend/Kventin.Services/Interfaces/Services/IRoleService.cs
@@ -4,10 +4,10 @@ namespace Kventin.Services.Interfaces.Services
 {
     public interface IRoleService
     {
-        public Task SetUserRoles(long authorizedUserId, long userId, List<string> rolenamesToAdd);
-        public Task DeleteUserRole(long authorizedUserId, long userId, List<string> rolenamesToDelete);
-        public Task<List<string>> GetUserRoles(long userId);
-        public Task<List<UserRoleInfoDto>> GetAllUsersWithRoles(long authorizedUserId);
+        public Task SetUserRoles(int authorizedUserId, int userId, List<string> rolenamesToAdd);
+        public Task DeleteUserRole(int authorizedUserId, int userId, List<string> rolenamesToDelete);
+        public Task<List<string>> GetUserRoles(int userId);
+        public Task<List<UserRoleInfoDto>> GetAllUsersWithRoles(int authorizedUserId);
         public Task<List<string>> GetAllRoles(List<string> authorizedUserRolenames);
     }
 }
