+++ Kventin.Backend/Kventin.Services/Interfaces/Services/IStudyGroupService.cs
@@ -5,11 +5,11 @@ namespace Kventin.Services.Interfaces.Services
     public interface IStudyGroupService
     {
         public Task CreateStudyGroup(CreateStudyGroupDto dto);
-        public Task DeleteStudyGroup(long studyGroupId);
-        public Task UpdateStudyGroup(long studyGroupId, UpdateStudyGroupDto dto);
-        public Task<StudyGroupFullInfoDto> GetStudyGroupFullInfo(long studyGroupId);
-        public Task<List<StudyGroupShortInfoDto>> GetAllStudyGroupsShortInfo(long userId, List<string> userRoles, long childId);
-        public Task AddStudentToStudyGroup(long studyGroupId, long studentId);
-        public Task DeleteStudentFromStudyGroup(long studyGroupId, long studentId);
+        public Task DeleteStudyGroup(int studyGroupId);
+        public Task UpdateStudyGroup(int studyGroupId, UpdateStudyGroupDto dto);
+        public Task<StudyGroupFullInfoDto> GetStudyGroupFullInfo(int studyGroupId);
+        public Task<List<StudyGroupShortInfoDto>> GetAllStudyGroupsShortInfo(int userId, List<string> userRoles, int childId);
+        public Task AddStudentToStudyGroup(int studyGroupId, int studentId);
+        public Task DeleteStudentFromStudyGroup(int studyGroupId, int studentId);
     }
 }
